with_langgraph/pyproject.toml
[project]
name = "with-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.19.0",
    "groq>=0.25.0",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "openai>=1.82.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
